SEED/requirements.txt
omegaconf
hydra-core
pyrootutils
timm
einops
sentencepiece
webdataset

# new
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
xformers==0.0.29.post2
peft==0.15.0
huggingface-hub==0.29.3
datasets==3.4.1
transformers==4.49.0
diffusers==0.32.2
trl==0.15.2
accelerate==1.5.2
sentence_transformers==4.0.2
vllm==0.8.2
ninja==1.11.1.3
wandb==0.19.8